# SelfLearning/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blog Q&A API Service",
    "description": "\n    API for blog content processing, question generation, and retrieval.\n    \n    ## Authentication\n    \n    This API uses two types of API key authentication:\n    \n    ### Admin Authentication (X-Admin-Key)\n    - Required for all `/api/v1/publishers/*` endpoints\n    - Used for publisher management and administration\n    - Header: `X-Admin-Key: admin_your-key-here`\n    \n    ### Publisher Authentication (X-API-Key)\n    - Required for content endpoints: `/api/v1/questions/by-url`, `/api/v1/jobs/process`\n    - Each publisher gets a unique API key upon onboarding\n    - Validates domain ownership automatically\n    - Header: `X-API-Key: pub_your-key-here`\n    \n    ## Features\n    - Automated blog content processing\n    - AI-generated questions and summaries\n    - Vector similarity search\n    - Custom Q&A endpoint\n    - Publisher management and rate limiting\n    ",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/publishers/onboard": {
      "post": {
        "tags": [
          "Publishers"
        ],
        "summary": "Onboard Publisher",
        "description": "Onboard a new publisher.\n\nCreates a new publisher account with custom configuration.\nReturns the publisher details including a generated API key.\n\nExample:\n```json\n{\n    \"name\": \"Tech Blog Inc\",\n    \"domain\": \"techblog.com\",\n    \"email\": \"admin@techblog.com\",\n    \"config\": {\n        \"questions_per_blog\": 7,\n        \"llm_model\": \"gpt-4o-mini\",\n        \"temperature\": 0.8,\n        \"daily_blog_limit\": 50\n    }\n}\n```",
        "operationId": "onboard_publisher_api_v1_publishers_onboard_post",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Admin API key for managing publishers",
              "title": "X-Admin-Key"
            },
            "description": "Admin API key for managing publishers"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublisherCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Publisher onboarded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherOnboardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Admin authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Publisher with domain already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/publishers/by-domain/{domain}": {
      "get": {
        "tags": [
          "Publishers"
        ],
        "summary": "Get Publisher By Domain",
        "description": "Get publisher by domain.\n\nUseful for checking if a domain is already onboarded.",
        "operationId": "get_publisher_by_domain_api_v1_publishers_by_domain__domain__get",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Admin API key for managing publishers",
              "title": "X-Admin-Key"
            },
            "description": "Admin API key for managing publishers"
          }
        ],
        "responses": {
          "200": {
            "description": "Publisher retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublisherGetResponse"
                }
              }
            }
          },
          "401": {
            "description": "Admin authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Publisher not found for domain",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publishers/": {
      "get": {
        "tags": [
          "Publishers"
        ],
        "summary": "List Publishers",
        "description": "List all publishers with pagination.\n\nCan filter by status (active, inactive, suspended, trial).",
        "operationId": "list_publishers_api_v1_publishers__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PublisherStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Admin API key for managing publishers",
              "title": "X-Admin-Key"
            },
            "description": "Admin API key for managing publishers"
          }
        ],
        "responses": {
          "200": {
            "description": "Publishers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishersListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Admin authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublisherCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublisherConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "subscription_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Tier",
            "default": "free"
          }
        },
        "type": "object",
        "required": [
          "name",
          "domain",
          "email"
        ],
        "title": "PublisherCreateRequest",
        "description": "Request model for creating a new publisher."
      },
...
      "PublisherStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "suspended",
          "trial"
        ],
        "title": "PublisherStatus",
        "description": "Publisher account status."
      },
...
    },
    "securitySchemes": {
      "AdminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key",
        "description": "Admin API key for publisher management endpoints"
      },
      "PublisherKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Publisher API key for content access and processing"
      }
    }
  }
}